[bos #45836] Adding the parametric update of the groups created by filters; modification of the UI for filters on group
- Adding the python dump of filters an ensuring its loading : the
content of the selection inside filters is now taken into account (no
empty selection)

- Adding the UI to remove filters and keep attributeSelectionList
objects : a disabled style is applied to the MultiSelector QListView
(list of attributes seen in a group) if filters are used

- Adapting the UI to the filters : if filters are used to create a
group, can not delete an element added in the list (right click - delete
action disabled) and can not add item from manual selection

- Taking into account retrocompatibility for scripts with no selection
in filter args : filters are removed when one of them has an empty
selection in a python script. A warning is sent in the Salome Python
console when the removal is made.

- Adding tests and doc for the "Remove filters" button

- Adding the CHECK_NAMING param to avoid bad selection by coords for
tests TestFilters_Mixed1 and Mixed2